Prep for refactoring allegations show
- Move card header to allegations edit view (it's already in the show)
- Card imports show container instead of old component
- Fix specs

[#refactor]
